rename some directories spiders works without alchemy
