+++ popcorn/spiders/base.py
@@ -2,5 +2,5 @@ from scrapy.spiders import CrawlSpider
 
 
 class BaseMixin(CrawlSpider):
-    def before_start(self, session):
+    def before_start(self):
         pass
