+++ popcorn_site/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for popcorn_page project.
+Django settings for popcorn_site project.
 
 Generated by 'django-admin startproject' using Django 2.1.7.
 
@@ -50,7 +50,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'popcorn_page.urls'
+ROOT_URLCONF = 'popcorn_site.urls'
 
 TEMPLATES = [
     {
@@ -68,7 +68,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'popcorn_page.wsgi.application'
+WSGI_APPLICATION = 'popcorn_site.wsgi.application'
 
 
 # Database
@@ -110,13 +110,13 @@ AUTH_PASSWORD_VALIDATORS = [
 
 LANGUAGE_CODE = 'en-us'
 
-TIME_ZONE = 'UTC'
+TIME_ZONE = 'Europe/Moscow'
 
 USE_I18N = True
 
 USE_L10N = True
 
-USE_TZ = True
+USE_TZ = False
 
 
 # Static files (CSS, JavaScript, Images)
