+++ popcorn_site/urls.py
@@ -1,4 +1,4 @@
-"""popcorn_page URL Configuration
+"""popcorn_site URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.1/topics/http/urls/
