+++ scrapy_app/management/commands/crawl_series.py
